requirements.txt
# Core LLM & Embeddings
openai>=1.30.0
sentence-transformers>=2.7.0

# Vector Database
faiss-cpu>=1.8.0

# API Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Data & Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
numpy>=1.26.0
tiktoken>=0.7.0

# Testing
pytest>=8.0.0
httpx>=0.27.0          
tqdm>=4.66.0
streamlit/